refactor(canvas): move chat shell adapters
