+++ apps/canvas-workspace/src/renderer/src/components/shell/ConversationCompletionToastBridge/index.test.tsx
@@ -2,15 +2,15 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it } from 'vitest';
-import { I18nProvider } from '../../i18n';
-import { AppShellProvider } from '../shell/AppShellProvider';
-import { conversationKey } from '../../../../shared/conversation-runtime';
+import { I18nProvider } from '../../../i18n';
+import { AppShellProvider } from '../AppShellProvider';
+import { conversationKey } from '../../../../../shared/conversation-runtime';
 import {
   recordConversationCompletion,
   resetConversationCompletionStoreForTests,
   setConversationVisible,
-} from './hooks/conversationCompletionStore';
-import { ConversationCompletionToastBridge } from './ConversationCompletionToastBridge';
+} from '../../chat/hooks/conversationCompletionStore';
+import { ConversationCompletionToastBridge } from './index';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
 let root: Root | null = null;
